requirements.txt
asgiref==3.8.1
boto3==1.40.24
botocore==1.40.24
Django==5.2.3
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
